requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
_tflow_select=2.2.0=eigen
abseil-cpp=20211102.0=hc377ac9_0
absl-py=1.4.0=py311hca03da5_0
aiohttp=3.8.5=py311h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
appnope=0.1.3=pyhd8ed1ab_0
arrow-cpp=11.0.0=he3f21e0_2
asttokens=2.4.0=pyhd8ed1ab_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py311hca03da5_0
attrs=23.1.0=py311hca03da5_0
aws-c-common=0.6.8=h80987f9_1
aws-c-event-stream=0.1.6=h313beb8_6
aws-checksums=0.1.11=h80987f9_2
aws-sdk-cpp=1.8.185=h4a942e0_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=openblas
blinker=1.4=py311hca03da5_0
boost-cpp=1.73.0=h1a28f6b_12
bottleneck=1.3.5=py311ha0d4635_0
branca=0.6.0=py311hca03da5_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py311h80987f9_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=h80987f9_0
ca-certificates=2023.08.22=hca03da5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.7.22=py311hca03da5_0
cffi=1.15.1=py311h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py311hca03da5_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.5=py311h48ca7d4_0
cryptography=41.0.3=py311h3c57c4d_0
cycler=0.11.0=pyhd3eb1b0_0
datetime=5.2=pypi_0
debugpy=1.6.7=py311h313beb8_0
decorator=5.1.1=pyhd8ed1ab_0
et_xmlfile=1.1.0=py311hca03da5_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.2.2=h04105a8_0
flatbuffers=2.0.0=hc377ac9_0
folium=0.14.0=py311hca03da5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py311h80987f9_0
gast=0.4.0=pyhd3eb1b0_0
geos=3.9.1=hc377ac9_1
gettext=0.21.0=h13f89a0_1
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_3
glog=0.5.0=hc377ac9_0
gmp=6.2.1=hc377ac9_3
gnutls=3.6.15=h887c41c_0
google-auth=2.22.0=py311hca03da5_0
google-auth-oauthlib=0.5.2=py311hca03da5_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.48.2=h877324c_0
grpcio=1.48.2=py311h877324c_0
h5py=3.9.0=py311hba6ad2f_0
hdf5=1.12.1=h160e8cb_2
icu=68.1=hc377ac9_0
idna=3.4=py311hca03da5_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.25.2=pyh1050b4e_0
ipython=8.15.0=pyh31c8845_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py311hca03da5_0
joblib=1.2.0=py311hca03da5_0
jpeg=9e=h80987f9_1
jsonschema=4.17.3=py311hca03da5_0
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=5.3.0=py311hca03da5_0
keras=2.12.0=py311hca03da5_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=py311h313beb8_0
krb5=1.20.1=h8380606_1
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libboost=1.73.0=h49e8a49_12
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcurl=8.2.1=h0f1d93c_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=hf27765b_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.4=h80987f9_0
libnghttp2=1.52.0=h10c0552_1
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.43.2=h091b4b1_0
libssh2=1.10.0=h449679c_2
libtasn1=4.19.0=h80987f9_0
libthrift=0.15.0=h169de6a_2
libtiff=4.5.1=h313beb8_0
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h372ba2a_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markdown=3.4.1=py311hca03da5_0
markupsafe=2.1.1=py311h80987f9_0
matplotlib-base=3.7.2=py311h7aedaa7_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
multidict=6.0.2=py311h80987f9_0
munkres=1.1.4=py_0
nbformat=5.9.2=py311hca03da5_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=h84b5d62_1
numexpr=2.8.4=py311h6dc990b_1
numpy=1.23.5=py311hb2c0538_0
numpy-base=1.23.5=py311h9eb1c70_0
oauthlib=3.2.2=py311hca03da5_0
openh264=1.8.0=h98b2900_0
openpyxl=3.0.10=py311h80987f9_0
openssl=1.1.1w=h1a28f6b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
orc=1.7.4=hdca1487_1
packaging=23.1=py311hca03da5_0
pandas=2.0.3=py311h7aedaa7_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311h313beb8_1
pip=23.2.1=py311hca03da5_0
platformdirs=3.10.0=py311hca03da5_0
plotly=5.9.0=py311hca03da5_0
polars=0.19.11=py311hcca7419_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=3.20.3=py311h313beb8_0
psutil=5.9.0=py311h80987f9_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py311h7575258_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pyhd8ed1ab_0
pyjwt=2.4.0=py311hca03da5_0
pyopenssl=23.2.0=py311hca03da5_0
pyparsing=3.0.9=py311hca03da5_0
pyrsistent=0.18.0=py311h80987f9_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.0=hf452327_1_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py311hca03da5_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.11=4_cp311
pytz=2023.3.post1=py311hca03da5_0
pyzmq=25.1.0=py311h313beb8_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
requests=2.31.0=py311hca03da5_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.2.2=py311h313beb8_1
scipy=1.11.1=py311hc76d9b0_0
seaborn=0.12.2=py311hca03da5_0
setuptools=68.0.0=py311hca03da5_0
shapely=2.0.1=py311h3713c0e_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=hc377ac9_0
sqlite=3.41.2=h80987f9_0
stack_data=0.6.2=pyhd8ed1ab_0
tenacity=8.2.2=py311hca03da5_0
tensorboard=2.12.1=py311hca03da5_0
tensorboard-data-server=0.7.0=py311ha6e5c4f_0
tensorboard-plugin-wit=1.8.1=py311hca03da5_0
tensorflow=2.12.0=eigen_py311h689d69b_0
tensorflow-base=2.12.0=eigen_py311h0a52ebb_0
tensorflow-estimator=2.12.0=py311hca03da5_0
termcolor=2.1.0=py311hca03da5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=hb8d0fd4_0
tornado=6.3.2=py311h80987f9_0
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.10.1=pyhd8ed1ab_0
typing_extensions=4.7.1=py311hca03da5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311hca03da5_0
utf8proc=2.6.1=h1a28f6b_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.2.3=py311hca03da5_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.14.1=py311h80987f9_0
x264=1!152.20180806=h1a28f6b_0
xz=5.4.2=h80987f9_0
yarl=1.8.1=py311h80987f9_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h53f4e23_5
zope-interface=6.0=pypi_0
zstd=1.5.5=hd90d995_0
